Fix throughput calculation.
DrawTPSChart() was not detecting cached stats results from the server. When a
cached result is returned, the calculated tps would be 0. Ignore cached
results.

Switch to using the timestamps in the stats results to calculate the duration
and the throughput, so that the function doesn't depend on the refresh frequency
to be 1 second. Even if it's 1 second, due to network delays and scheduling, it
could be more or less than 1 second.